Fix build on android_aarch64.
Whether UCHAR_MAX expands to a signed or unsigned number appears to be a
matter of some debate. Or the Android headers are wrong. Just add a cast
and not think about it too hard.

Reviewed-on: https://boringssl-review.googlesource.com/15385
Commit-Queue: David Benjamin <davidben@google.com>

+++ crypto/compiler_test.cc
@@ -58,7 +58,7 @@ static void CheckRepresentation(T value) {
 
 TEST(CompilerTest, IntegerRepresentation) {
   EXPECT_EQ(8, CHAR_BIT);
-  EXPECT_EQ(0xff, UCHAR_MAX);
+  EXPECT_EQ(0xff, static_cast<int>(UCHAR_MAX));
 
   // uint8_t is assumed to be unsigned char. I.e., casting to uint8_t should be
   // as good as unsigned char for strict aliasing purposes.
